Wechar_Develop/feature2-ai/backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.4
